bcachefs: Copygc fixup.
allowed_fragmentation will sometimes be 0 on full systems. Even though
fragmentation is almost never 0 due to remaining data being smaller than
the bucket size, this would result in waiting on 0 IO before starting up
again. When allowed_fragmentation is 0, we now set it to bucket size so
the system sleeps until we have enough data to fill the bucket.

Signed-off-by: Daniel Hill <daniel@gluo.nz>